Update and bump to 0.2.0.0

+++ TheDialgaTeam.Xiropht.Xirorig.Benchmark/GenerateNumberMathCalculationBenchmark.cs
@@ -12,7 +12,7 @@ namespace TheDialgaTeam.Xiropht.Xirorig.Benchmark
         private readonly byte[] _randomNumber = new byte[1];
 
         [Benchmark]
-        public decimal GenerateNumberMathCalculation()
+        public float GenerateNumberMathCalculation()
         {
             return MiningUtility.GenerateNumberMathCalculation(_rngCryptoServiceProvider, _randomNumber, 2, 1000000, 1, 7);
         }
